Report of current disk use for videos, auto-tagging for cup sizes and more
Yapo e+ now reports current video disk use on startup. Later, it will be added to the video scanner. The use is shown in MB, GB or TB depending on the amount of data.

Also added a bra cup routine which tags the actor with a tag for the bra cup, ex "D Cup". In addition, it tags for a group of breast sizes, like "Big tits".

Remember to run pip install --upgrade -r requirements.txt as I just upgraded the requirements for Django and other modules.