Moved scrapers to local module
Moved the scrapers to local module, modified the Freeones scraper slightly.
Added a check which displays various statistics in console on startup.

The system is now prepared for filename-based scene and release recognition, which will be useful for scraping for release information. At this point, it will only recognice videos from Facial Fest (based on a filename convention of FFxxxx where x are digits, which should match the release IDs and official naming), and only in the log.