Introduced exclusions to scene/actor tags and websites
It is now possible to exclude certain words so the auto-tagging doesn't tag certain words (fx the word "floral" in the filename would trigger a tag named "oral" - by excluding "floral" you can prevent this).
Another example is the fact that there is a website named "Stepsiblings Caught" and another one just named "Stepsiblings". By adding "Stepsiblings" in the exclusion list for "Stepsiblings Caught", the system will not incorrectly register the website "Stepsiblings" for a file that should only be registered to "Stepsiblings Caught".
Exclusion lists are available for both scene and actor tags plus websites.