Fixed error, date omitted from title renamer
The title renamer had a small error that resulted in a possibility for the date not to be added to a new title, if rename has been turned on.
Introduced Google and Bing image searching utils and a smart photo cropper (not requiring OpenCV)
Work is being done to auto-scrape tag images.
YAPO now reports version number on navbar (if I can remember to change it)